refactor(skills): relocate skills module to codemie-code agent
- Move src/skills/ to src/agents/codemie-code/skills/
- Update import paths in affected files
- Remove obsolete documentation files (HOOKS-ITERATION-IMPLEMENTATION.md, SKILLS_INTEGRATION_TEST_REPORT.md)
